fix: introducing @Lazy activities
